Rename geometry module to vectors
